Separate rules for common/content/glossary
